LayoutTests/imported/w3c/web-platform-tests/css/css-display/display-first-line-001-expected.txt

First line is float and flex.
First line is float but not flex.
First line is flex but not float.
First line is not float or flex.

FAIL display of first-line assert_equals: expected "inline" but got "flex"
